+++ src/database/operations/baseFields/index.ts
@@ -1,3 +1,3 @@
 export * from './createOrUpdateBaseField';
 export * from './loadBaseField';
-export * from './loadBaseFields';
+export * from './loadBaseFieldBundle';